Refactor build scripts
